Decompose Backend into BackendBase + subsystem mixins
Split the ~1400-line monolithic `Backend` class in `backend/state.py` into a
shared `BackendBase` kernel plus coupling-aligned subsystem mixins under a new
`backend/ops/` subpackage. `state.py` becomes a slim assembler composing the
mixins into `Backend`.

- `ops/base.py` (BackendBase): instance state, virtual clock/snowflake helpers,
  gateway `emit`, universal getters, and `compute_permissions`.
- Coupling-aligned mixins: tightly-coupled methods share a mixin (guilds +
  members + roles; channels + threads + stage instances; messages + auto-mod),
  so every cross-mixin call targets only BackendBase.
- Serializer params retyped `backend: Backend` -> `backend: BackendBase` (every
  member they touch lives on the base), so a mixin `self` type-checks.

Pure structural refactor: every method body moved verbatim and the `Backend`
public surface is byte-identical (verified). Mixin names follow a uniform
`<Domain>Mixin` scheme; the design rationale lives in `ops/__init__.py`.
