+++ changes/+backend-decomposition.misc.md
@@ -1,0 +1,1 @@
+Split the monolithic `Backend` class into a shared `BackendBase` kernel plus coupling-aligned subsystem mixins under `simcord.backend.ops`. Pure structural refactor — the `Backend` public surface and all behaviour are unchanged.
